requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
colorama==0.4.4
peewee==3.14.10
peewee-db-evolve==3.7.6
psycopg2==2.9.3
python-decouple==3.6
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.1
